requirements.txt
langgraph
openai
gitpython
python-dotenv
tiktoken
streamlit
langchain
langchain-community
langchain-openai